scripts/requirements_import.txt
libsql-client==0.7.0
pandas==2.1.4
python-dotenv==1.0.0
